Create default Layout component
